# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FarmEasy API - Indoor farm cycle management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation by token (public, no auth — invitee may have no account yet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
      responses:
        "201":
          description: Invitation accepted, membership created
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizationId:
                    type: integer
                  role:
                    type: string
                  email:
                    type: string
                required: [organizationId, role, email]
        "400":
          description: Invalid, expired, or already-used invitation; or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /members/{userId}/role:
    patch:
      operationId: changeMemberRole
      tags: [members]
      summary: Change a member's role (admin/technician only, owner is immutable)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "404":
          description: Member not found or not modifiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /members/{userId}:
    delete:
      operationId: removeMember
      tags: [members]
      summary: Soft-remove a member (owner is immutable)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "404":
          description: Member not found or not removable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AcceptInvitationRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token]
...
    ChangeRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, technician]
      required: [role]