# lib/api-zod/src/generated/types/changeRoleRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FarmEasy API - Indoor farm cycle management
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeRoleRequestRole } from './changeRoleRequestRole';

export interface ChangeRoleRequest {
  role: ChangeRoleRequestRole;
}